fix restore if database does not exists

+++ clanModules/postgresql/default.nix
@@ -146,6 +146,7 @@ in
             pkgs.coreutils
             pkgs.util-linux
             pkgs.zstd
+            pkgs.gnugrep
           ]
         }
         while [[ "$(systemctl is-active postgresql)" == activating ]]; do
@@ -167,7 +168,9 @@ in
             }
 
             mkdir -p "${folder}"
-            runuser -u postgres -- dropdb "${db.name}"
+            if runuser -u postgres -- psql -d postgres -c "SELECT 1 FROM pg_database WHERE datname = '${db.name}'" | grep -q 1; then
+              runuser -u postgres -- dropdb "${db.name}"
+            fi
             runuser -u postgres -- pg_restore -C -d postgres "${current}"
           )
         else
